% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ast_turnover_ratio.R
\name{ast_turnover_ratio}
\alias{ast_turnover_ratio}
\title{Calculate the assist to turnover ratio of a player from the Smith WBB dataset.}
\arguments{
\item{input_player}{Player from the Smith WBB roster to calculate stat for.}

\item{input_season}{Specifies season within \code{input_player}s career to calculate stat for.}
}
\value{
The assist to turnover ratio of \code{input_player} in \code{input_season}.
}
\description{
Calculate the assist to turnover ratio of a player from the Smith WBB dataset.
}
\examples{
ast_turnover_ratio(input_player = "Bondi, Lauren", input_season = "2018-19")
# [1] 1.27381
}
