% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform_players.R
\name{perform_players}
\alias{perform_players}
\title{Compare field goals attempted of players in one season}
\usage{
perform_players(player_num1, player_num2, input_season)
}
\arguments{
\item{player_num1}{The number of the first player of interest}

\item{player_num2}{The number of the second player of interest}

\item{input_season}{The season of interest from the dataset}
}
\value{
A bar graph that compares performance of \code{player1}'s performance with \code{player2}'s on field goals attempted in one \code{season}
}
\description{
Compare field goals attempted of players in one season
}
\examples{
perform_players(10, 22, "2021-22")
}
