% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{smith_oppwbb_data}
\alias{smith_oppwbb_data}
\title{2010-2022 Cumulative Statistics for Opponents of Smith College Women Basketball Game}
\format{
\code{smith_oppwbb_data} is an object of class tbl_df (inherits from tbl, data.frame) with 153 rows and 25 columns
\describe{
\item{opponent}{The opponent against Smith College}
\item{date}{The date of the game}
\item{score}{Smith's score vs. opponent's score}
\item{w_l}{Whether Smith won or lost}
\item{fgm}{Field Goals Made}
\item{fga}{Field Goals Attempted}
\item{fg_pct}{Field Goal Percentage}
\item{x3fgm}{Three Point Field Goals Made}
\item{x3fga}{Three Point Field Goals Attempted}
\item{x3fg_pct}{Three Point Field Goal Percentage}
\item{ftm}{Free Throws Made}
\item{fta}{Free Throws Attempted}
\item{ft_pct}{Free Throw Percentage}
\item{off}{Offensive Rebounds}
\item{def}{Defensive Rebounds}
\item{tot}{Total Rebounds}
\item{avg}{Average Rebounds Per Game}
\item{pf}{Personal Fouls}
\item{ast}{Assists}
\item{to}{Turnovers}
\item{stl}{Steals}
\item{blk}{Blocks}
\item{pts}{Total Points Scored}
\item{avg_2}{Average Points Scored Per Game}
\item{season}{Season Year}
}
}
\source{
\code{smith_oppwbb_data} comes from the \href{https://smithpioneers.com/sports/womens-basketball/stats/2010-11}{Smith College Pioneers Basketball Statistics page}
}
\usage{
smith_oppwbb_data
}
\description{
\code{smith_oppwbb_data} contains data from smith college women basketball games
}
\examples{
smith_oppwbb_data
}
\keyword{datasets}
