% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{smith_wbb_data}
\alias{smith_wbb_data}
\title{2010-2022 Cumulative Statistics for Smith College Women Basketball Game}
\format{
\code{smith_wbb_data} is an object of class tbl_df (inherits from tbl, data.frame) with 152 rows and 27 columns
\describe{
\item{number}{Player's Jersey Number}
\item{player}{Player's Name}
\item{gp}{Games Played}
\item{gs}{Games Started}
\item{tot}{Total Minutes Played}
\item{avg}{Average Minutes Played Per Game}
\item{fgm}{Field Goals Made}
\item{fga}{Field Goals Attempted}
\item{fg_percent}{Field Goal Percentage}
\item{x3pt}{Three Pointers Made}
\item{x3pta}{Three Pointers Attempted}
\item{x3pt_percent}{Three Pointer Percentage}
\item{ftm}{Free Throws Made}
\item{fta}{Free Throws Attempted}
\item{ft_percent}{Free Throw Percentage}
\item{pts}{Total Points Scored}
\item{avg2}{Average Points Scored Per Game}
\item{off}{Offensive Rebounds}
\item{def}{Defensive Rebounds}
\item{tot_2}{Total Rebounds}
\item{avg_3}{Average Rebounds Per Game}
\item{pf}{Personal Fouls}
\item{ast}{Assists}
\item{to}{Turnovers}
\item{stl}{Steals}
\item{blk}{Blocks}
\item{season}{Season Year}
}
\source{
\code{smith_wbb_data} comes from the \href{https://smithpioneers.com/sports/womens-basketball/stats/2010-11}{Smith College Pioneers Basketball Statistics page}
}
\usage{
smith_wbb_data
}
\description{
\code{smith_wbb_data} contains data from smith college women basketball games
}
\examples{
smith_wbb_data
}
\keyword{datasets}
