% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/offensive_rating.R
\name{offensive_rating}
\alias{offensive_rating}
\title{Calculate Offensive Rating from the Smith WBB dataset.}
\arguments{
\item{input_player}{Player from the Smith WBB roster to calculate stat for.}

\item{input_season}{Specifies season within \code{input_player}s career to calculate stat for.}
}
\value{
The offensive rating of \code{input_player} in \code{input_season}.
}
\description{
Calculate Offensive Rating from the Smith WBB dataset.
}
\examples{
offensive_rating(input_player = "Bondi, Lauren", input_season = "2018-19")
# [1]
player <- "Dornbusch, Abby"
season <- "2010-11"
offensive_rating(player, season)
# [1]
}
