% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateefficiency.R
\name{TS_percentage}
\alias{TS_percentage}
\title{Calculate TS\% or True Shooting Percentage from the Smith WBB dataset.}
\usage{
TS_percentage(input_player, input_season)
}
\arguments{
\item{input_player}{Player from the Smith WBB roster to calculate stat for.}

\item{input_season}{Specifies season within \code{input_player}s career to calculate stat for.}
}
\value{
The true shooting percentage of \code{input_player} in \code{input_season}.
}
\description{
Calculate TS\% or True Shooting Percentage from the Smith WBB dataset.
}
\examples{
#TS_percentage(player = "Bondi, Lauren", season = "2018-19")
#input_player <- "Dornbusch, Abby"
#input_season <- "2010-11"
#TS_percentage(input_player, input_season)
}
